Include package-level failures in formatted test output
In some cases, it's possible for a package to be marked as a test
failure even if no tests inside it have failed. The motivating example
for this change is a timeout: a test overshooting the alotted timeout
is considered by `go test` to be a package-level failure, even if no
tests _inside_ the package are consdered failures.

This led to cases where the user woudl see an "All tests passed" message
from the go test filter, but still mysteriously fail the make step.

To address this, the test renderer now
 * treats package-level pass/fail/skip events as first-class citizens
   and includes them in its event count,
 * tracks the cahced test output at both a package and individual test
   level, end
 * displays the whole package output if a package is marked as failed,
   but only if there is no obvious failed test top account for the
   package-level failure.

This patch also removes the json files created by the unit tests, as
they are not yet needed for anything.