Fix non-JSON fandling in test reporter
A non-json line in a test log caused the test reporter to enter an
infinite loop. This is caused by the JSON fragment reader attempting to
pick itself up after a failure, when instead it really should just
signal en error and exit.

Unfortunately, `go test -json` does not report build failures in the JSON
format as all of its other messages, but dumps a plain text logline into
the output.

This patch makes the test renderer exit on a parse failure, rather than
try to carry on.