Resolve race in db tests
When disconnecting a client due to an expired certificate, the monitor
was emitting the disconnection event before forcing the client to
disconnect. This led to races in tests that wait on the disonnection
event, where tests would:

 1. wait for the disconnection event before proceeding
 2. receive the disconnection event
 2. take action based on the disconnection event, assuming that the
    client was disconnected, when there is no guarantee that the
    disconnection has actually happened yet.

This patch re-orders the disconnection and event broadcast, such that
the disconnection happens _first_, meaning that if a watcher receives
the doisconnection event, then the dosconnection has already been
attempted.