Fix app server goroutine leak
Fixes https://github.com/gravitational/teleport-private/issues/79
LAT-APP21-4: DOS - Goroutine leak in app server

Change app server listener.Accept method to return an error the second
time it is called. This will trigger the http.Serve function to exit
gracefully rather than block until the server is closed.

This also changes the checkHTTPResponse server test. Closing the server
in the test no longer produces a context canceled error because that
error was coming from the blocked listener when the server called
listener.Close.

This change uses the new ServerConnState type in httplib that tracks
connection state for an http.Server instance. This is used to wait for
the connection to close before returning from HandleConnection.