Skip unit & integration tests if no code changes
Skips long-running tests if only doc changes are detected in a PR. Uses
a Go script to compare the changes between the target branch and the
supplied PR code. If those changes are only to files deemed to be docs
(i.e. docs/**, **.md or **.mdx), then the script writes a marker file
that later build steps can use to determin whether to run tests or not.

The whole marker-filr approach is largely due to GCB not allowing
optional steps or having the notion of returning success early. Each
test needs to independently check if it should run or not, and an easy
way to do that is to check if the marker exists.