[v6.2] Skip tests on a docs-only PR (#9416) (#9512)
Skips long-running tests if only doc changes are detected in a PR. Uses
a Go script to compare the changes between the target branch and the
supplied PR code. If those changes are only to files deemed to be docs
(i.e. docs/**, rfd/**, **.md or **.mdx), then the script will finish early and return
success.

If the script detects non-doc changes it will go on to build and run the
appropriate tests.