Added stateless merge queue.
Added stateless merge queue that uses a cron workflow to loop over all
Pull Requests every 15 minutes. If the Pull Request has auto merge
enabled, it will attempt to update the PR if it has gone out of date
with the base branch.

This mimics a real merge queue, but does it in a stateless manner.