Recommend a highly available LDAP endpoint.
Early versions of our docs were worded in such a way that suggested
pointing Teleport at a single domain controller, which would result
in issues if a DC goes down for maintenance or due to error.

We also used to make insecure LDAP connections on port 389, and then
upgrade them using a STARTTLS operation. This is no longer supported.

Fix both of these issues by recommending a highly available LDAP
server and removing references to port 389.