docs: recommend a highly available LDAP endpoint. (#9744)
* Recommend a highly available LDAP endpoint.

Early versions of our docs were worded in such a way that suggested
pointing Teleport at a single domain controller, which would result
in issues if a DC goes down for maintenance or due to error.

We also used to make insecure LDAP connections on port 389, and then
upgrade them using a STARTTLS operation. This is no longer supported.

Fix both of these issues by recommending a highly available LDAP
server and removing references to port 389.

* Add note about direct vs IoT mode

* Move the desktop config reference to a shared includes file

This will prevent the desktop access config examples from diverging
in different parts of the docs.

Additionally, simplify the example in the getting started guide,
as the goal of that guide is to get up and running quickly, not
to provide an exhaustive reference of all the configuration options.