Remove TestTwoClustersTunnel integration test.
TestTwoClustersTunnel has legitimately been failing for many months.
However, we've been re-running integration tests until it passes.

On master and branch/v8 we've fixed the issue by upgrading etcd and
grpc dependency.

https://github.com/gravitational/teleport/pull/#9656
https://github.com/gravitational/teleport/pull/#9607

Ideally we'd fix TestTwoClustersTunnel on branch/6.2 and branch/v7 by
updating etcd and grpc, but as these branches are meant to be fairly
stable at this point, and the risk of unintended consequences by
updating etcd and grpc feels too high.