Fix Doctests
The linting applied to file via the tools in next appears to vary if the files
are outside the /src directory: If the files are under /src, a rigorous lint
is applied. If the files are outside of /src, a "less-rigorous" lint is
applied, letting many legitimate issues slip through.

This patch alters the CI script to symlink the /workspace directory (the
mount-point that GCB uses to inject code into the container running a build
step) under the next image's /src/content directory, so that the correct,
rigorous lint will be applied.

It also fixes the lint errors that have crept in during the time the linter
was being incorrectly lenient.

See-Also: #9600
See-Also: #10107