Fixed TestDisconnection
Integration tests were creating the users SSH and x509 certificates at
the time of server creation, specifically at the time of the
"TeleInstance" creation.

If any component within the "TeleInstance" was slow to start, for example
if the proxy watcher had to be reset as can be seen below, the users
certificate would often be expired by the time the SSH request was issued
which would lead to the SSH server rejecting the connection and the test
failing with an unexpected error.

For most tests this is not an issue, because test certificates are valid
for 24 hours. However "TestDisconnection" scopes the TTL of certificates
down (often times to about 2 seconds) to test different certificate
lifetime disconnection scenarios which would lead to this test being flaky.

Updated integration test logic to instead "issue" the users certificate at
the time of client creation instead of server creation.