Add a conceptual guide to Moderated Sessions (#10094)
The tutorial-style guide created in previous commits did not
provide enough context into how the feature works and which
configuration options are available. Since the instructions in
the guide are mostly example configurations anyway, I thought it
would make sense to combine the guide and the reference while
removing the tutorial instructions and adding more detailed
guidance on available configuration options.

This guide now lives in docs/pages/access-controls. I thought
keeping it in docs/pages/setup/reference would not be appropriate,
since the references there are for broad approaches to interacting
with Teleport (configuration, dynamic resources, metrics, etc.),
rather than specific features.

We can plan to edit the dynamic resource reference to include
the config options mentioned in this guide.