Tag build images with teleport8 instead of go version (#10211)
Prior to this patch, the buildbox images used for CI have always just been tagged with the version of Go they support. In the past this has been discrete enough to stay inside individual release boundaries.

Recently, though, the version of Go used by Teleport 8 and Teleport 9 has remained constant, while other requirements of the the build environment have changed, leading to the unfortunate situation where publishing a buildbox image that works for Teleport 9 breaks the Teleport 8 build.

This patch de-couples the version of the buildbox from the version of Go installed on it, and produces a new teleport8-tagged buildbox image for Teleport 8 builds to use.