Use an apt-key alternative in install instructions
Backports #10084

* Use an apt-key alternative in install instructions

Current installation instructions for Debian/Ubuntu advise using
apt-key, which is deprecated. This change uses the recommended
method of appending a "signed-by" argument to the APT sources
list along with the rest of a repo's information.

Also creates a partial with Linux installation info to avoid
having to repeat changes in our different Linux installation
instructions.

* Respond to PR feedback

- Add "sudo" before the installation script
- Add command prompts before commands (not part of PR feedback but
  something I noticed was missing)