Use yamlv3 to parse Teleport config file
V2 of the YAML library incorrectly decoded child section keys to a
map[interface{}]interface{} resulting in an error when converting to a
struct using utils.ObjectToStruct. This prevented extending the storage
section of the configuration file to support child sections.

V3 fixed this issue by decoding to map[string]interface{} when all
child keys are strings. This change allows us to extend the storage
section to include child sections.

See godoc in the included test for details.